+++ src/frontend/components/VideoPlayer/index.tsx
@@ -16,7 +16,6 @@ import {
   videoSize,
 } from '../../types/tracks';
 import { VideoPlayerInterface } from '../../types/VideoPlayer';
-import { isMSESupported } from '../../utils/isMSESupported';
 import { Maybe, Nullable } from '../../utils/types';
 import { DownloadVideo } from '../DownloadVideo';
 import { ERROR_COMPONENT_ROUTE } from '../ErrorComponent/route';
@@ -131,14 +130,6 @@ const VideoPlayer = ({
         poster={thumbnailUrls[resolutions[resolutions.length - 1]]}
         tabIndex={-1}
       >
-        {isMSESupported() && (
-          <source
-            src={video.urls.manifests.hls}
-            size="auto"
-            type="application/x-mpegURL"
-          />
-        )}
-
         {resolutions.map((size) => (
           <source
             key={video.urls.mp4[size]}